Fix bug where patterns would not be detected
Fixed a bug in the SiddhiEngine where patterns would not be
detected if one sensor is used multiple times in the same pattern.

Found a new bug in the Socket handling:
If a client is listening for socket events and two messages are sent
at the same time, the handler crashes.