Implement in-memory DB to improve tests
Implemented an in-memory DB that allows to test the server without
needing an actual Database. This enables tests that do not depend
on the DB state.
Technically, the DBConnectorMock should be tested as well... maybe
in the next phase.

Adjusted the SiddhiEngineTest to use the new DB implementation.
Made minor changes to ServerInstance that were necessary to be able
to create ServerInstances that use the new DB implementation.