Exclude pattern id from handler tag
excluded the source pattern id from the handler tag as there is
currently no way to determine the pattern id of the output node
from the input node. The tag will be updated as soon as this
was fixed.

Added TODOs at the relevant places.

Proof of concept test case was implemented and is currently ignored.